chore(concerns): Rename traits to concerns
